+++ app/screens/code-of-conduct/code-of-conduct-screen.tsx
@@ -1,11 +1,12 @@
 import * as React from "react"
-import { TextStyle, ViewStyle, Image, View } from "react-native"
-import { Text } from "../../components/text"
+import { Image, TextStyle, View, ViewStyle } from "react-native"
 import { NavigationScreenProps } from "react-navigation"
-import { Screen } from "../../components/screen"
-import { palette } from "../../theme/palette"
-import { color, spacing } from "../../theme"
+import { BackButton } from "../../components/back-button"
 import { Contact } from "../../components/contact"
+import { Screen } from "../../components/screen"
+import { Text } from "../../components/text"
+import { color, spacing } from "../../theme"
+import { palette } from "../../theme/palette"
 
 export interface CodeOfConductScreenProps extends NavigationScreenProps<{}> {}
 
@@ -13,24 +14,12 @@ const ROOT: ViewStyle = {
   marginHorizontal: spacing.large,
   paddingBottom: spacing.huge,
 }
-const HEADER_TEXT: TextStyle = {
-  fontSize: 17,
-  fontWeight: "600",
-  lineHeight: 45,
-  letterSpacing: 3.0,
-  marginLeft: spacing.large,
-  color: color.palette.shamrock,
-}
+
 const TITLE: TextStyle = {
   marginTop: spacing.extraLarge,
   letterSpacing: 1.68,
 }
-const BACK_ARROW: ViewStyle = {
-  flexDirection: "row",
-  paddingLeft: spacing.large,
-  alignItems: "center",
-  justifyContent: "center",
-}
+
 const SECTION: TextStyle = {
   marginVertical: spacing.medium,
 }
@@ -43,20 +32,6 @@ const SECTION_TITLE: TextStyle = {
   marginTop: spacing.medium,
 }
 
-const HIT_SLOP = {
-  top: 20,
-  left: 20,
-  right: 20,
-  bottom: 20,
-}
-
-const backArrow = backTitle => (
-  <View style={BACK_ARROW} hitSlop={HIT_SLOP}>
-    <Image source={require("../../components/title-bar/icon.back-arrow.png")} />
-    <Text text={backTitle} style={HEADER_TEXT} />
-  </View>
-)
-
 const email = "conf@infinite.red"
 const twitter = "chainreactconf"
 const phoneNumber = "(360) 562-0450"
@@ -64,7 +39,7 @@ const phoneNumber = "(360) 562-0450"
 export class CodeOfConductScreen extends React.Component<CodeOfConductScreenProps, {}> {
   static navigationOptions = ({ navigation, _ }) => ({
     headerStyle: { backgroundColor: color.palette.portGore, borderBottomWidth: 0 },
-    headerBackImage: backArrow(navigation.getParam("backTitle", "INFO")),
+    headerBackImage: <BackButton backTitle={navigation.getParam("backTitle", "INFO")} />,
     headerTintColor: color.palette.shamrock,
   })
