+++ storybook/storybook-registry.ts
@@ -23,3 +23,4 @@ require("../app/components/contact/contact.story")
 require("../app/components/photobomb/photobomb.story")
 require("../app/components/blog-link/blog-link.story")
 require("../app/components/bullet-item/bullet-item.story")
+require("../app/components/back-button/back-button.story")
